+++ CHANGELOG.md
@@ -14,9 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.1.0] - 2019-06-25
 ### Added
 - Expose `_root` property in queried instances to allow identify the root instance.
+- Emit `_root` property on cleanAny events.
 
 ### Fixed
 - Expose custom queries methods in `customQueries` property, as described in documentation.
+- Expose test.queries properties for concurrent sources.
 
 ## [1.0.0] - 2019-06-03
 ### BREAKING CHANGES